+++ src/crawlee/browsers/_playwright_browser_controller.py
@@ -29,8 +29,8 @@ logger = getLogger(__name__)
 class PlaywrightBrowserController(BaseBrowserController):
     """Controller for managing Playwright browser instances and their pages.
 
-    This class provides methods to manage pages within a browser instance, ensuring that the number
-    of open pages does not exceed the specified limit and tracking the state of the pages.
+    It provides methods to control browser instances, manage their pages, and handle context-specific
+    configurations. It enforces limits on the number of open pages and tracks their state.
     """
 
     AUTOMATION_LIBRARY = 'playwright'
@@ -171,7 +171,7 @@ class PlaywrightBrowserController(BaseBrowserController):
         )
 
         if proxy_info:
-            if browser_new_context_options['proxy']:
+            if browser_new_context_options.get('proxy'):
                 logger.warning("browser_new_context_options['proxy'] overriden by explicit `proxy_info` argument.")
 
             browser_new_context_options['proxy'] = ProxySettings(